requirements.txt
biplist==1.0.3
certifi==2022.12.7
charset-normalizer==2.1.1
idna==3.4
pycryptodome==3.16.0
requests==2.28.1
urllib3==1.26.13
